added cart , orders pages and updated blogs,consults,products admin pages

+++ frontend/src/components/types/ConsultationTableColumns.tsx
@@ -60,13 +60,17 @@ export const columns: ColumnDef<Consultation>[] = [
             }
 
             const handleDelete = async ()=>{
-                try{
-                    const response = await axios.delete(`http://localhost:8000/api/v1/consults/${consultation._id}`,{withCredentials:true});
-                    if(response.data.status === "success"){
-                        window.location.reload();
-                    }    
-                }catch(error){
-                    console.log('error deleting this consultation', error);
+                const confirmDelete = window.confirm('Are you sure you want to delete this consultation?');
+
+                if (confirmDelete) {
+                    try {
+                        const response = await axios.delete(`http://localhost:8000/api/v1/consults/${consultation._id}`, { withCredentials: true });
+                        if (response.data.status === "success") {
+                            window.location.reload();
+                        }
+                    } catch (error) {
+                        console.log('Error deleting this consultation', error);
+                    }
                 }
             }
